Split automation folder in 2
