+++ src/main/java/automationSamples/AccountBalanceAutomation.java
@@ -1,5 +1,7 @@
-package automations;
+package automationSamples;
 
+import automationTools.QueryManager;
+import automationTools.AbstractPeopleSoftAutomation;
 import io.CsvFileRequirements;
 import java.util.Arrays;
 import org.openqa.selenium.By;
