+++ src/main/java/automationTools/ReadingAutomation.java
@@ -1,4 +1,4 @@
-package automations;
+package automationTools;
 
 /**
  *
