+++ src/main/java/gui/AutomationInfoBox.java
@@ -1,6 +1,6 @@
 package gui;
 
-import automations.AbstractAutomation;
+import automationTools.AbstractAutomation;
 import java.awt.BorderLayout;
 import javax.swing.JLabel;
 import javax.swing.JPanel;
